+++ core/src/com/mygdx/panzer/Settings.java
@@ -7,7 +7,7 @@ import com.badlogic.gdx.math.Vector2;
  */
 
  public final class Settings {
-    public static final float SPEED_LIMIT = 10;
+    public static final int SPEED_LIMIT = 10;
     public static final int SPEED_MIN = 1;
     public static final int SENSORS_LIMIT = 500;
     public static final int SENSORS_MIN = 10;
@@ -16,7 +16,7 @@ import com.badlogic.gdx.math.Vector2;
     private static Vector2 startPos = new Vector2(0, 0);
     private static Vector2 finishPos = new Vector2(800, 600);
     private static int sensorRange = 100;
-    private static float maxSpeed = 3;
+    private static int maxSpeed = 3;
     private static int startAngle = 0;
     public static final int WORLD_WIDTH = 2048;
     public static final int WORLD_HEIGHT = 1152;
@@ -38,11 +38,11 @@ import com.badlogic.gdx.math.Vector2;
         startAngle = newRotation;
     }
 
-    static public float getMaxSpeed() {
+    static public int getMaxSpeed() {
         return maxSpeed;
     }
 
-    static public void setMaxSpeed(float newMaxSpeed) {
+    static public void setMaxSpeed(int newMaxSpeed) {
         maxSpeed = newMaxSpeed;
     }
